parse7: Prep to move rules into new Parser class.
I think the rules will now be later in the code as I want to set them up as
directly as possible as members of the Parser instance. I'm also, for now,
including a comment-only description of the grammar as it will help me both
document and that the grammar makes sense to me.