Fix some code/grammar bugs.
These iterations are slowly getting us to the point of successfully parsing
04.input. I modified mode popping to happen on a specific seq-rule item with the
value '<pop>' (no single quotes in the actual grammar). I now remember an old
design question that came up, which is how to deal with mode parses that fail. I
don't completely remember, but I vaguely think that I might be able to have a
'<fail>' parse item that indicates a parse failure once you hit it. Or to simply
let a parse fail in the standard way, and let that propagate up out of an
implicit mode pop. Yea, that last one sounds good. Although we may eventually
find use for a <fail> special item.